+++ TombLib/TombLib.csproj
@@ -141,6 +141,7 @@
     <Compile Include="GeometryIO\RoomGeometryExporter.cs" />
     <Compile Include="Graphics\Frustum.cs" />
     <Compile Include="LevelData\Compilers\Util\BatchCompileList.cs" />
+    <Compile Include="LevelData\Instances\ObjectGroup.cs" />
     <Compile Include="LevelData\Instances\ScriptInstance.cs" />
     <Compile Include="LevelData\Instances\TriggerVolumeInstance.cs" />
     <Compile Include="NG\NgEncryption.cs" />
